changes to allow new AuthorizerData implementations for StandardAuthorizer
